pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evalmedia"
version = "0.1.1"
description = "Open-source framework for evaluating AI-generated media quality"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "evalmedia contributors" },
]
keywords = ["ai", "media", "evaluation", "quality", "image", "agents", "vlm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
    "Typing :: Typed",
]
dependencies = [
    "Pillow>=10.0",
    "httpx>=0.25",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
claude = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
classical = ["torch>=2.0", "open-clip-torch>=2.24"]
cli = ["typer>=0.12", "rich>=13.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]
all = ["evalmedia[claude,openai,classical,cli]"]

[project.scripts]
evalmedia = "evalmedia.cli.main:app"

[project.urls]
Homepage = "https://evalmedia.com"
Repository = "https://github.com/saidkaban/evalmedia"
Issues = "https://github.com/saidkaban/evalmedia/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/evalmedia"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["evalmedia"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
